AI/requirements.txt
tensorflow
numpy
fastapi
uvicorn
pydantic
pillow
aiofiles
python-multipart
requests
